requirements.txt
tweepy>=4.14.0
python-dotenv>=1.0.1
Unidecode>=1.3.8
